use makeDecorator
