+++ .storybook/preview.js
@@ -1,0 +1,5 @@
+// import { customUserAgents } from "./userAgent";
+
+export const parameters = {
+  //   userAgent: customUserAgents,
+};
